軽くリファクタリング
