+++ app/api/ai-analysis/route.ts
@@ -1,6 +1,6 @@
+import { ai } from "@/src/lib/gemini/gemini";
 import { CommentItem } from "@/src/tools/demand-analysis/types/youtubeApiTypes";
 import { fetchData } from "@/src/tools/demand-analysis/utils/fetchData";
-import { GoogleGenAI } from "@google/genai";
 import { NextRequest, NextResponse } from "next/server";
 
 export async function POST(req: NextRequest) {
@@ -18,7 +18,6 @@ export async function POST(req: NextRequest) {
         { status: 404 },
       );
 
-    const ai = new GoogleGenAI({ apiKey: process.env.GEMINI_API_KEY });
     const thumbnail = await fetch(thumbnailUrl);
     const imageArrayBuffer = await thumbnail.arrayBuffer();
     const base64ImageData = Buffer.from(imageArrayBuffer).toString("base64");
